update config section set
